Made corrections to what was reviewed
