+++ server/src/modules/community-event/router.ts
@@ -1,7 +1,6 @@
 import { Router } from 'express';
 import communityEventController from './controller';
 
-
 /**
  * Router for the Community Event module
  */
@@ -9,6 +8,6 @@ const communityEventRouter = Router();
 
 communityEventRouter.post('/', communityEventController.create);
 
-communityEventRouter.put('/', communityEventController.updateCommunityEvent);
+communityEventRouter.put('/', communityEventController.update);
 
 export default communityEventRouter;
