--- server/src/modules/community-event/validations/index.ts
@@ -1,1 +1,0 @@
-export * from "./updateCommunityEventReqVal"
