--- server/src/utils/errors/index.ts
@@ -1,6 +1,0 @@
-export * from "./appError"
-export * from "./internalServerError"
-export * from "./notFoundError"
-export * from "./unauthorizedError"
-export * from "./conflictError"
-export * from "./badRequestError"
