--- server/src/utils/errors/unauthorizedError.ts
@@ -1,11 +1,0 @@
-import { ApplicationError } from './appError';
-
-
-/**
- * Unauthorized Error: request requires user authentication
- */
-export class UnauthorizedError extends ApplicationError {
-  constructor(message?: string) {
-    super(message || 'UNAUTHORIZED', 401);
-  }
-}
